+++ threedi_graph.py
@@ -2,8 +2,8 @@
 """
 /***************************************************************************
  ThreeDiToolbox
-                                 A QGIS plugin
- Toolbox for working with 3di hydraulic models
+                                 A QGIS plugin for working with 3di
+                                 hydraulic models
                               -------------------
         begin                : 2016-03-04
         git sha              : $Format:%H$
@@ -23,8 +23,6 @@
 import os.path
 from PyQt4.QtCore import Qt
 import qgis
-
-from utils.user_messages import log, pop_up_info
 
 from views.graph import GraphDockWidget
 
@@ -63,8 +61,7 @@ class ThreeDiGraph:
         #self.ts_datasource.close()
         #self.ts_datasource = None
 
-    # noinspection PyPep8Naming
-    def onCloseChildWidget(self, widget_nr):
+    def on_close_child_widget(self, widget_nr):
         """Cleanup necessary items here when plugin dockwidget is closed"""
         nr = None
 
@@ -77,23 +74,21 @@ class ThreeDiGraph:
         # close widget
         if nr is not None:
             widget = self.dock_widgets[nr]
-            widget.closingWidget.disconnect(self.onCloseChildWidget)
+            widget.closingWidget.disconnect(self.on_close_child_widget)
 
             del self.dock_widgets[nr]
 
     def run(self):
         """
         Run method that loads and starts the plugin (docked graph widget)
-        :return:
         """
-
         # create the dockwidget
         self.widget_nr += 1
         new_widget = GraphDockWidget(self.iface, parent_class=self, nr=self.widget_nr, ts_datasource=self.ts_datasource)
         self.dock_widgets.append(new_widget)
 
         # connect cleanup on closing of dockwidget
-        new_widget.closingWidget.connect(self.onCloseChildWidget)
+        new_widget.closingWidget.connect(self.on_close_child_widget)
 
         # show the dockwidget
         self.iface.addDockWidget(Qt.BottomDockWidgetArea, new_widget)
@@ -104,6 +99,3 @@ class ThreeDiGraph:
             window.tabifyDockWidget(self.dock_widgets[0], new_widget)
 
         new_widget.show()
-
-
-