+++ utils/user_messages.py
@@ -2,7 +2,7 @@
 
 from PyQt4.QtGui  import QMessageBox
 from qgis.core import QgsMessageLog
-
+from qgis.utils import iface
 
 def log(msg, level='INFO'):
     """Shortcut for QgsMessageLog.logMessage function."""
@@ -14,3 +14,7 @@ def log(msg, level='INFO'):
 def pop_up_info(msg='', title='Information', parent=None):
     """Display an info message via Qt box"""
     QMessageBox.information(parent, title, '%s' % msg)
+
+def statusbar_message(msg=''):
+    """Display message in status bar """
+    iface.mainWindow().statusBar().showMessage(msg)