+++ CHANGES.rst
@@ -4,7 +4,11 @@ threedi-qgis-plugin changelog
 0.3 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Add tool version number to about box.
+- Add support of multiple result files.
+- Warn user on adding to many locations to graph.
+- Only new locations will be added to graph.
+- Improved color selections for timeseries after the first 20.
 
 
 0.2 (2016-04-12)